% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy.R
\name{add_dummy}
\alias{add_dummy}
\title{Additional \link[base]{logical} Predictor}
\usage{
add_dummy(
  start.model,
  x,
  data = eval(start.model$call$data),
  mc.cores = getOption("cores"),
  ...
)
}
\arguments{
\item{start.model}{a regression model, e.g.,
\link[stats]{lm}, \link[stats]{glm}, or \link[survival]{coxph}, etc.}

\item{x}{one-sided \link[stats]{formula},
\link[base]{numeric} predictors \eqn{x}'s as the columns of one \link[base]{matrix} column in \code{data}}

\item{data}{(optional) \link[base]{data.frame} in the model \link[base]{call} of \code{start.model}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}
}
\value{
The function \code{\link[=add_dummy]{add_dummy()}} returns an object of class \code{'add_dummy'}.
}
\description{
Additional \link[base]{logical} Predictor
}
\keyword{internal}
