% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy_partition.R
\name{add_dummy_partition}
\alias{add_dummy_partition}
\title{Additional \link[base]{logical} Predictor via Repeated Partitioning}
\usage{
add_dummy_partition(
  start.model,
  x,
  data = eval(start.model$call$data),
  times,
  mc.cores = getOption("cores"),
  ...
)
}
\arguments{
\item{start.model}{a regression model, e.g.,
\link[stats]{lm}, \link[stats]{glm}, or \link[survival]{coxph}, etc.}

\item{x}{one-sided \link[stats]{formula},
\link[base]{numeric} predictors \eqn{x}'s as the columns of one \link[base]{matrix} column in \code{data}}

\item{data}{(optional) \link[base]{data.frame} in the model \link[base]{call} of \code{start.model}}

\item{times, ...}{additional parameters of function \link[caret]{createDataPartition} or \code{\link[=statusPartition]{statusPartition()}}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}
}
\value{
The function \code{\link[=add_dummy_partition]{add_dummy_partition()}} returns an object of \link[base]{class} \code{'add_dummy'}.
}
\description{
Additional \link[base]{logical} Predictor via Repeated Partitioning
}
\keyword{internal}
