% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy_rSplit.R
\name{add_dummy_rSplit}
\alias{add_dummy_rSplit}
\title{Dichotomizing Predictors via Repeated Sample Splits}
\usage{
add_dummy_rSplit(
  start.model,
  x,
  data = eval(start.model$call$data),
  n,
  mc.cores = switch(.Platform$OS.type, windows = 1L, detectCores()),
  ...
)
}
\arguments{
\item{start.model}{a regression model, e.g.,
\link[stats]{lm}, \link[stats]{glm}, or \link[survival]{coxph}, etc.}

\item{x}{one-sided \link[stats]{formula},
\link[base]{numeric} predictors \eqn{x}'s as the columns of one \link[base]{matrix} column in \code{data}}

\item{data}{(optional) \link[base]{data.frame} in the model \link[base]{call} of \code{start.model}}

\item{n, ...}{additional parameters of function \link{rSplit} for function \link{add_dummy_rSplit}.
For function \code{add_dummy}, these parameters are not in use}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}
}
\value{
Function \link{add_dummy_rSplit} returns an object of \link[base]{class} \code{'add_dummy_rSplit'}, which is a \link[base]{list} of dichotomizing \link[base]{function}s.
}
\description{
Dichotomizing predictors using repeated sample splits.
}
\details{
Function \link{add_dummy_rSplit} dichotomizes predictors via repeated sample splits. Specifically,

\enumerate{
\item Generate multiple, i.e., repeated, training-test sample splits (via \link{rSplit})
\item For each candidate predictor \eqn{x_i}, find the \emph{\strong{median-split-dichotomized regression model}} based on the repeated sample splits, see functions \link{splitd_} and \link{quantile.splitd.list};
}
}
\examples{
library(spatstat.grouped)
library(spatstat.grouped.data)
library(survival) # to help ?spatstat.geom::hyperframe understand ?survival::Surv

s = grouped_ppp(hladr + phenotype ~ OS + gender + age | patient_id/image_id, data = wrobel_lung)
sQ = s |>
 aggregate_quantile(by = ~ patient_id, probs = seq.int(from = 0, to = 1, by = .1))
dim(sQ)

sQ0 = sQ[1:100,] # training set
sQ1 = sQ[-(1:100),] # test set

set.seed(2364); m1 = coxph(OS ~ gender, data = sQ0) |>
 add_dummy_rSplit(x = ~ hladr.quantile, n = 20L) |> subset(subset = p1 > .15 & p1 < .85) |>
 sort_by(y = abs(cf)) |>
 head(n = 2L)
m1
#hladr.quantile[, "30\%"]>=0.15409
#hladr.quantile[, "40\%"]>=0.21358
predict(m1, newdata = sQ1)

}
