% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_num.R
\name{add_num}
\alias{add_num}
\title{add_num}
\usage{
add_num(
  start.model,
  x,
  data = eval(start.model$call$data),
  mc.cores = switch(.Platform$OS.type, windows = 1L, detectCores()),
  ...
)
}
\arguments{
\item{start.model}{a regression model (e.g., \link[stats]{lm}, \link[stats]{glm}, or \link[survival]{coxph}, etc.)}

\item{x}{one-sided \link[stats]{formula} to specify
the \link[base]{numeric} predictors \eqn{x}'s as the columns of one \link[base]{matrix} column in \code{data}}

\item{data}{\link[base]{data.frame}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}

\item{...}{additional parameters, currently of no use}
}
\value{
Function \code{\link[=add_num]{add_num()}} ..
}
\description{
add_num
}
\details{
Function \code{\link[=add_num]{add_num()}} ..
}
\examples{
# vignette('intro', package = 'maxEff')
}
