% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numeric.R
\name{add_numeric}
\alias{add_numeric}
\title{Additional Predictor as \link[base]{numeric}}
\usage{
add_numeric(
  start.model,
  x,
  data = eval(start.model$call$data),
  mc.cores = getOption("cores"),
  ...
)
}
\arguments{
\item{start.model}{a regression model (e.g., \link[stats]{lm}, \link[stats]{glm}, or \link[survival]{coxph}, etc.)}

\item{x}{one-sided \link[stats]{formula} to specify
the \link[base]{numeric} predictors \eqn{x}'s as the columns of one \link[base]{matrix} column in \code{data}}

\item{data}{\link[base]{data.frame}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}

\item{...}{additional parameters, currently of no use}
}
\value{
The function \code{\link[=add_numeric]{add_numeric()}} returns an \link{add_numeric} object,
which is a \link[stats]{listof} objects with an internal class \code{'add_numeric_'}.
}
\description{
Additional predictor as \link[base]{numeric}.
}
\details{
The function \code{\link[=add_numeric]{add_numeric()}} treats each additional predictor as a \link[base]{numeric} variable,
and \link[stats]{update}s the starting model with each additional predictor.
}
\keyword{internal}
