% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node1.R
\name{get_cutoff}
\alias{get_cutoff}
\alias{get_cutoff.node1}
\title{Get Cutoff Value from a Dichotomizing Rule \code{\link[=node1]{node1()}}}
\usage{
get_cutoff(x)

\method{get_cutoff}{node1}(x)
}
\arguments{
\item{x}{see Usage}
}
\value{
The \code{S3} method \code{\link[=get_cutoff.node1]{get_cutoff.node1()}} returns a \link[base]{numeric} scalar.
}
\description{
To get the cutoff value from a Dichotomizing Rule \code{\link[=node1]{node1()}}.
}
\keyword{internal}
