% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dummy.R
\name{head.split_dummy}
\alias{head.split_dummy}
\title{head.split_dummy}
\usage{
\method{head}{split_dummy}(x, n, ...)
}
\arguments{
\item{x}{a \link{split_dummy} object}

\item{n}{positive \link[base]{integer} scalar}
}
\value{
Function \link{head.split_dummy} returns a \link{split_dummy} object.
}
\description{
head.split_dummy
}
\keyword{internal}
