% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy.R
\name{labels.add_dummy}
\alias{labels.add_dummy}
\title{\link{labels.add_dummy}}
\usage{
\method{labels}{add_dummy}(object, ...)
}
\arguments{
\item{object}{a \link{add_dummy} object}

\item{...}{..}
}
\value{
The \code{S3} method \code{\link[=labels.add_dummy]{labels.add_dummy()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
\link{labels.add_dummy}
}
\keyword{internal}
