% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numeric.R
\name{labels.add_numeric}
\alias{labels.add_numeric}
\title{\link{labels.add_numeric}}
\usage{
\method{labels}{add_numeric}(object, ...)
}
\arguments{
\item{object}{a \link{add_numeric} object}

\item{...}{..}
}
\value{
The \code{S3} method \code{\link[=labels.add_numeric]{labels.add_numeric()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
\link{labels.add_numeric}
}
\keyword{internal}
