% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node1.R
\name{labels.node1}
\alias{labels.node1}
\title{Find \link[base]{labels} from \link{node1}}
\usage{
\method{labels}{node1}(object, ...)
}
\arguments{
\item{object}{a \link{node1} object}

\item{...}{additional parameters, currently not in use}
}
\value{
The \code{S3} method \code{\link[=labels.node1]{labels.node1()}} returns a \link[base]{character} scalar.
}
\description{
Find \link[base]{labels} from \link{node1}
}
\keyword{internal}
