% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{maxEff-package}
\alias{maxEff}
\alias{maxEff-package}
\title{maxEff: Additional Predictor with Maximum Effect Size}
\description{
Methods of selecting one from many numeric predictors for a regression model, to ensure that the additional predictor has the maximum effect size.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tingtingzhan/maxEff}
  \item \url{https://tingtingzhan.quarto.pub/groupedhyperframe/maxEff.html}
  \item \url{https://tingtingzhan-groupedhyperframe.netlify.app/maxEff.html}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
\keyword{internal}
