% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy.R
\name{predict.add_dummy}
\alias{predict.add_dummy}
\title{Regression Models with Optimal Dichotomizing Predictors}
\usage{
\method{predict}{add_dummy}(object, ...)
}
\arguments{
\item{object}{an \link{add_dummy} object}

\item{...}{additional parameters of function \link{predict.splitd}, e.g., \code{newdata}}
}
\value{
Function \link{predict.add_dummy} returns a \link[base]{list} of regression models.
}
\description{
Regression models with optimal dichotomizing predictor(s), used either as boolean or continuous predictor(s).
}
\examples{
# see ?add_dummy
}
