% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dummy.R
\name{predict.split_dummy}
\alias{predict.split_dummy}
\title{Regression Models with Optimal Dichotomizing Predictors}
\usage{
\method{predict}{split_dummy}(object, newdata = attr(object, which = "data", exact = TRUE), ...)
}
\arguments{
\item{object}{an \link{split_dummy} object}

\item{newdata}{(optional) test \link[base]{data.frame}, candidate \link[base]{numeric} predictors \eqn{x}'s must have the same \link[base]{name} and \link[base]{dim}ension as the training data. If missing, the training data is used}

\item{...}{additional parameters of function \link{predict.splitd}}
}
\value{
Function \link{predict.split_dummy} returns a \link[base]{list} of regression models, \link[survival]{coxph} model for \link[survival]{Surv} response, \link[stats]{glm} for \link[base]{logical} response, and \link[stats]{lm} model for \link[base]{numeric} response.
}
\description{
Regression models with optimal dichotomizing predictor(s), used either as boolean or continuous predictor(s).
}
\examples{
# see ?`Qindex-package`
}
