% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitd.R
\name{predict.splitd}
\alias{predict.splitd}
\title{Regression Models with Optimal Dichotomizing Predictors}
\usage{
\method{predict}{splitd}(
  object,
  newdata,
  formula = attr(object, which = "formula", exact = TRUE),
  ...
)
}
\arguments{
\item{object}{an \link{splitd} object}

\item{newdata}{\link[base]{data.frame}, candidate \link[base]{numeric} predictors \eqn{x}'s must have the same \link[base]{name} and \link[base]{dim}ension as the training data. If missing, the training data is used}

\item{formula}{..}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{predict.splitd} returns a regression model,
either a \link[survival]{coxph} model for \link[survival]{Surv} response,
a \link[stats]{glm} for \link[base]{logical} response,
or a \link[stats]{lm} model for \link[base]{numeric} response.
}
\description{
Regression models with optimal dichotomizing predictor(s), used either as boolean or continuous predictor(s).
}
\keyword{internal}
