% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy.R
\name{predict_add_dummy}
\alias{predict_add_dummy}
\alias{predict.add_dummy}
\alias{predict.add_dummy_}
\title{Regression Models with Optimal Dichotomizing Predictors}
\usage{
\method{predict}{add_dummy}(object, ...)

\method{predict}{add_dummy_}(object, newdata, ...)
}
\arguments{
\item{object}{an \code{add_dummy_} object, as an element of the \link[stats]{listof} return from functions \code{\link[=add_dummy]{add_dummy()}} or \code{\link[=add_dummy_partition]{add_dummy_partition()}}}

\item{...}{additional parameters, currently not in use}

\item{newdata}{\link[base]{data.frame}, candidate \link[base]{numeric} predictors \eqn{x}'s must have the same \link[base]{name} and \link[base]{dim}ension as the training data. If missing, the training data is used}
}
\value{
The \code{S3} method \code{\link[=predict.add_dummy]{predict.add_dummy()}} returns a \link[stats]{listof} regression models.

The function \code{\link[=predict.add_dummy_]{predict.add_dummy_()}} returns a updated regression model.
}
\description{
Regression models with optimal dichotomizing predictor(s), used either as boolean or continuous predictor(s).
}
\keyword{internal}
