% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numeric.R
\name{predict_add_numeric}
\alias{predict_add_numeric}
\alias{predict.add_numeric}
\alias{predict.add_numeric_}
\title{Regression Models with Optimal Dichotomizing Predictors}
\usage{
\method{predict}{add_numeric}(object, ...)

\method{predict}{add_numeric_}(object, newdata, ...)
}
\arguments{
\item{object}{an \link{add_numeric} object}

\item{...}{additional parameters of function \code{predict.add_numeric_}, e.g., \code{newdata}}
}
\value{
The \code{S3} method \code{\link[=predict.add_numeric]{predict.add_numeric()}} returns a \link[stats]{listof} regression models.
}
\description{
Regression models with optimal dichotomizing predictor(s), used either as boolean or continuous predictor(s).
}
\keyword{internal}
