% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_S3.R
\name{print.add_}
\alias{print.add_}
\title{print.add_}
\usage{
\method{print}{add_}(x, ...)
}
\arguments{
\item{x}{an \link{add_dummy} or \link{add_num} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{print.add_} does not have a returned value
}
\description{
print.add_
}
\details{
..
}
\keyword{internal}
