% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy.R
\name{print.add_dummy}
\alias{print.add_dummy}
\title{\link{print.add_dummy}}
\usage{
\method{print}{add_dummy}(x, ...)
}
\arguments{
\item{x}{an object returned from functions
\code{\link[=add_dummy_partition]{add_dummy_partition()}} or \code{\link[=add_dummy]{add_dummy()}}}

\item{...}{additional parameters, currently not in use}
}
\value{
The \code{S3} method \code{\link[=print.add_dummy]{print.add_dummy()}} does not have a returned value
}
\description{
\link{print.add_dummy}
}
\details{
..
}
\keyword{internal}
