% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numeric.R
\name{print.add_numeric}
\alias{print.add_numeric}
\title{\link{print.add_numeric}}
\usage{
\method{print}{add_numeric}(x, ...)
}
\arguments{
\item{x}{a \link{add_numeric} object}

\item{...}{..}
}
\description{
\link{print.add_numeric}
}
\keyword{internal}
