% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dummy.R
\name{print.split_dummy}
\alias{print.split_dummy}
\title{print.split_dummy}
\usage{
\method{print}{split_dummy}(x, ...)
}
\arguments{
\item{x}{a \link{split_dummy} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{print.split_dummy} does not have a returned value
}
\description{
print.split_dummy
}
\details{
..
}
\keyword{internal}
