% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitd.R
\name{quantile.splitd.list}
\alias{quantile.splitd.list}
\title{Quantile of Split-Dichotomized Regression Models}
\usage{
\method{quantile}{splitd.list}(x, probs, ...)
}
\arguments{
\item{x}{a \code{'splitd.list'} object, returned from function \link{.splitd_}}

\item{probs}{\link[base]{double} scalar or \link[base]{vector}, see \link[stats]{quantile}}
}
\value{
Function \link{quantile.splitd.list} returns a \code{'splitd.list'} object.
}
\description{
Quantile(s) of \code{'splitd.list'} object.
}
\details{
Function \link{quantile.splitd.list} is a method dispatch of the S3 generic function \link[stats]{quantile} on \code{'splitd.list'} object.
Specifically,

\enumerate{
\item {collect the univariable regression coefficient estimate from each \link{.splitd} model;}
\item {find the nearest-even (i.e., \code{type = 3}) \link[stats]{quantile} of the coefficients from Step 1;}
\item {the \link{.splitd} models corresponding to the selected coefficient quantile(s) in Step 2, is returned.}
}
}
\keyword{internal}
