% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpartD.R
\name{rpartD_}
\alias{rpartD_}
\title{Batch Operation of \link{rpartD}}
\usage{
rpartD_(
  y,
  X,
  mc.cores = switch(.Platform$OS.type, windows = 1L, detectCores()),
  ...
)
}
\arguments{
\item{y, ...}{see function \link{rpartD}}

\item{X}{\link[base]{data.frame} or \link[base]{list}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}
}
\value{
Function \link{rpartD_} returns a \link[base]{list} of \link{rpartD} returns.
}
\description{
Batch Operation of \link{rpartD}
}
