% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_S3.R
\name{sort_by.add_}
\alias{sort_by.add_}
\title{\link{sort_by.add_}}
\usage{
\method{sort_by}{add_}(x, y, ...)
}
\arguments{
\item{x}{an object returned from functions
\code{\link[=add_dummy_partition]{add_dummy_partition()}}, \code{\link[=add_dummy]{add_dummy()}} or \code{\link[=add_numeric]{add_numeric()}}}

\item{y}{\link[base]{language}, see function \link[base]{sort_by}}

\item{...}{additional parameters of S3 generic \link[base]{sort_by}, etc.}
}
\value{
The \code{S3} method \code{\link[=sort_by.add_]{sort_by.add_()}} returns an object of the same \link[base]{class} as input \code{x}.
}
\description{
\link{sort_by.add_}
}
\details{
The \code{S3} method \code{\link[=sort_by.add_]{sort_by.add_()}} sorts the elements of an \code{'add_'} object by a certain criterion \code{y}.
We suggest using \code{y = abc(effsize)} and \code{decreasing = TRUE} order of the \link[base]{abs}olute values of the effect sizes of additional predictor.
}
\keyword{internal}
