% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dummy.R
\name{sort_by.split_dummy}
\alias{sort_by.split_dummy}
\title{\link{sort_by.split_dummy}}
\usage{
\method{sort_by}{split_dummy}(x, y, ...)
}
\arguments{
\item{x}{\link{split_dummy} object}

\item{y}{\link[base]{language}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{sort_by.split_dummy} returns a \link{split_dummy} object.
}
\description{
..
}
\details{
Sort the elements of a \link{split_dummy} object by a certain criterion \code{y}.

We suggest using \code{y = abc(cf)}, indicating a \code{decreasing} order of the \link[base]{abs}olute values of the regression coefficient estimate of the median-split-dichotomized regression models.
}
\keyword{internal}
