% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedPartition.R
\name{stratifiedPartition}
\alias{stratifiedPartition}
\title{Stratified Partition}
\usage{
stratifiedPartition(y, times, p = 0.8, ...)
}
\arguments{
\item{y}{response \eqn{y}, a \link[base]{double},
\link[base]{logical} or \link[base]{factor} \link[base]{vector},
or a \link[survival]{Surv} object}

\item{times}{positive \link[base]{integer} scalar \eqn{n}, number of \link[base]{replicate}s of partitions}

\item{p}{\link[base]{double} scalar between 0 and 1,
percentage \eqn{p} of training subjects, default \code{.8}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=stratifiedPartition]{stratifiedPartition()}} returns a length-\eqn{n} \link[stats]{listof}
\link[base]{integer} \link[base]{vector}s.
In each \link[base]{integer} \link[base]{vector} indicates the training subjects.
}
\description{
A partition, stratified based on the type of the response.
}
\details{
See \code{vignette('intro', package = 'maxEff')}.
}
\note{
Nomenclature of the parameters, as well as the returned object, of function \code{\link[=stratifiedPartition]{stratifiedPartition()}} follows those of function \link[caret]{createDataPartition}.

Function \code{caTools::sample.split} is not what we need.
}
\examples{
# vignette('intro', package = 'maxEff')
}
\keyword{internal}
