% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy.R
\name{subset.add_dummy}
\alias{subset.add_dummy}
\title{S3 Method Dispatches to \code{'add_dummy'} Class}
\usage{
\method{subset}{add_dummy}(x, subset, ...)
}
\arguments{
\item{x}{an object returned from functions \code{\link[=add_dummy_partition]{add_dummy_partition()}} or \code{\link[=add_dummy]{add_dummy()}}}

\item{subset}{\link[base]{language}}

\item{...}{additional parameters of function \code{\link[=predict.add_dummy_]{predict.add_dummy_()}}, e.g., \code{newdata}}
}
\value{
The \code{S3} method \code{\link[=subset.add_dummy]{subset.add_dummy()}} returns a \code{\link[=add_dummy]{add_dummy()}} object.
}
\description{
S3 Method Dispatches to \code{'add_dummy'} Class
}
\details{
The \code{S3} method \code{\link[=subset.add_dummy]{subset.add_dummy()}}, default subset \code{(p1>.15 & p1<.85)}.
See explanation of \eqn{p_1} in function \code{\link[=splitd]{splitd()}}.
}
\keyword{internal}
