% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy_rSplit.R
\name{subset.add_dummy_rSplit}
\alias{subset.add_dummy_rSplit}
\title{subset.add_dummy_rSplit}
\usage{
\method{subset}{add_dummy_rSplit}(x, subset, ...)
}
\arguments{
\item{x}{a \link{add_dummy_rSplit} object}

\item{subset}{\link[base]{language}}
}
\value{
Function \link{subset.add_dummy_rSplit} returns a \link{add_dummy_rSplit} object.
}
\description{
subset.add_dummy_rSplit
}
\details{
Default \code{(p1>.15 & p1<.85)}.
See explanation of \eqn{p_1} in function \link{splitd}.
}
\keyword{internal}
