% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dummy.R
\name{subset.split_dummy}
\alias{subset.split_dummy}
\title{subset.split_dummy}
\usage{
\method{subset}{split_dummy}(x, subset, ...)
}
\arguments{
\item{x}{a \link{split_dummy} object}

\item{subset}{\link[base]{language}}
}
\value{
Function \link{subset.split_dummy} returns a \link{split_dummy} object.
}
\description{
subset.split_dummy
}
\details{
Default \code{(p1>.15 & p1<.85)}.
See explanation of \eqn{p_1} in function \link{splitd}.
}
\keyword{internal}
