% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart1.R
\name{get_cutoff}
\alias{get_cutoff}
\alias{get_cutoff.rpart1}
\title{Get Cutoff Value from a Dichotomizing Rule \link{rpart1}}
\usage{
get_cutoff(x)

\method{get_cutoff}{rpart1}(x)
}
\arguments{
\item{x}{see Usage}
}
\value{
Function \link{get_cutoff.rpart1} returns a \link[base]{numeric} scalar.
}
\description{
To get the cutoff value from a Dichotomizing Rule \link{rpart1}.
}
