% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numeric.R
\name{labels.add_numeric}
\alias{labels.add_numeric}
\title{\link{labels.add_numeric}}
\usage{
\method{labels}{add_numeric}(x, ...)
}
\arguments{
\item{x}{a \link{add_numeric} object}

\item{...}{..}
}
\description{
\link{labels.add_numeric}
}
\keyword{internal}
