% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{maxEff-package}
\alias{maxEff}
\alias{maxEff-package}
\title{maxEff: Additional Predictor with Maximum Effect Size}
\description{
Methods of selecting one from many numeric predictors for a regression model, to ensure that the additional predictor has the maximum effect size.
}
\note{
Help files of individual functions are intentionally suppressed in this \code{pdf} manual.
Users are encouraged to get started with

\code{vignette('intro', package = 'maxEff')}
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} (\href{https://orcid.org/0000-0002-9104-4505}{ORCID})
}

}
