% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node1.R
\name{node1}
\alias{node1}
\title{Dichotomize via 1st Node of Recursive Partitioning}
\usage{
node1(x, ...)
}
\arguments{
\item{x}{an \link[rpart]{rpart.object}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=node1]{node1()}} returns an object of class \code{'node1'},
which is a \link[base]{function}
with one parameter \code{newx} taking a \link[base]{double} \link[base]{vector}.
}
\description{
Dichotomize via 1st Node of Recursive Partitioning
}
\note{
In future \link[base]{integer} and \link[base]{factor} predictors will be supported.

Function \link[rpart]{rpart} is quite slow.
}
\examples{
# See intro vignette, section Appendix, subsection node1()
}
\keyword{internal}
