% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_num.R
\name{predict.add_num}
\alias{predict.add_num}
\title{Regression Models with Optimal Dichotomizing Predictors}
\usage{
\method{predict}{add_num}(object, ...)
}
\arguments{
\item{object}{an \link{add_num} object}

\item{...}{additional parameters of function \code{predict.add_num_}, e.g., \code{newdata}}
}
\value{
Function \code{\link[=predict.add_num]{predict.add_num()}} returns a \link[stats]{listof} regression models.
}
\description{
Regression models with optimal dichotomizing predictor(s), used either as boolean or continuous predictor(s).
}
\keyword{internal}
