% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node1.R
\name{predict.node1}
\alias{predict.node1}
\title{Predict by \code{\link[=node1]{node1()}}}
\usage{
\method{predict}{node1}(object, newdata, ...)
}
\arguments{
\item{object}{a \link{node1} object}

\item{newdata}{a \link[base]{data.frame} or \link[spatstat.geom]{hyperframe}}

\item{...}{place holder for \code{S3} generic}
}
\description{
Predict by \code{\link[=node1]{node1()}}
}
\keyword{internal}
