% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_S3.R
\name{sort_by.add_}
\alias{sort_by.add_}
\title{\link{sort_by.add_}}
\usage{
\method{sort_by}{add_}(x, y, ...)
}
\arguments{
\item{x}{an \link{add_dummy_rSplit} or \link{add_num} object}

\item{y}{\link[base]{language}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=sort_by.add_]{sort_by.add_()}} returns an \link{add_dummy_rSplit} or \link{add_num} object.
}
\description{
..
}
\details{
Sort the elements of an \link{add_dummy_rSplit} or \link{add_num} object by a certain criterion \code{y}.

We suggest using \code{y = abc(effsize)}, indicating a \code{decreasing} order of the \link[base]{abs}olute values of the regression coefficient estimate of the median-split-dichotomized regression models.
}
\keyword{internal}
