% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy.R
\name{subset.add_dummy}
\alias{subset.add_dummy}
\title{subset.add_dummy}
\usage{
\method{subset}{add_dummy}(x, subset, ...)
}
\arguments{
\item{x}{a \link{add_dummy_rSplit} and/or \link{add_dummy} object}

\item{subset}{\link[base]{language}}
}
\value{
Function \code{\link[=subset.add_dummy]{subset.add_dummy()}} returns a \code{\link[=add_dummy]{add_dummy()}} object.
}
\description{
subset.add_dummy
}
\details{
Default \code{(p1>.15 & p1<.85)}.
See explanation of \eqn{p_1} in function \code{\link[=splitd]{splitd()}}.
}
\keyword{internal}
